initial work on combining mosaic-related content sections
